fix: create upgrader package
