+++ go.mod
@@ -4,7 +4,7 @@ go 1.17
 
 require (
 	github.com/bwmarrin/discordgo v0.23.2
-	github.com/google/go-github/v41 v41.0.0
+	github.com/google/go-github/v41 v41.0.1-0.20211227215900-a899e0fadbec
 	github.com/sirupsen/logrus v1.8.1
 	github.com/slack-go/slack v0.10.0
 	github.com/spf13/viper v1.9.0
